+++ packages/f-checkout/src/components/Address.vue
@@ -11,6 +11,7 @@
                 name="address-line-1"
                 :label-text="$t('labels.line1')"
                 label-style="inline"
+                is-grouped
                 :has-error="isAddressLine1Empty">
                 <template #error>
                     <error-message
@@ -27,6 +28,7 @@
                 :class="$style['c-address-formField']"
                 name="address-line-2"
                 :label-text="$t('labels.line2')"
+                is-grouped
                 label-style="inline" />
         </fieldset>
 
@@ -138,14 +140,12 @@ $address-weight-bold     : $font-weight-bold;
 }
 
 .c-address-group {
-    margin: spacing(x2) 0 spacing(x4);
+    margin: spacing(x2) 0;
     padding: 0;
     border: none;
     @include font-size($address-fontSize);
 
     .c-address-formField {
-        margin-bottom: -17px;
-
         &:focus-within,
         &:active {
             z-index: zIndex(high);
@@ -155,6 +155,6 @@ $address-weight-bold     : $font-weight-bold;
 }
 
 .c-address-error {
-    margin-bottom: spacing(x5);
+    margin-bottom: spacing(x2);
 }
 </style>