+++ packages/f-checkout/test/specs/component/f-checkout-collection.component.spec.js
@@ -3,8 +3,13 @@ import DemoControls from '../../../test-utils/component-objects/demo-controls';
 import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
 
 describe('f-checkout "collection" component tests', () => {
-    forEach(['mobileNumber'])
-    .it('each fields error message should be displayed', field => {
+      beforeEach(() => {
+        browser.url('?path=/story/components-organisms--checkout-component');
+        browser.switchToFrame(0);
+        CheckoutComponent.waitForCheckoutComponent();
+    });
+
+    it.skip('should display "mobileNumber" error message when collection method is set', () => {
         // Arrange
         DemoControls.selectCheckoutMethod('collection');
 
@@ -12,11 +17,11 @@ describe('f-checkout "collection" component tests', () => {
         CheckoutComponent.submit();
 
         // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed(field)).toBe(true);
+        expect(CheckoutComponent.isFieldErrorDisplayed('mobileNumber')).toBe(true);
     });
 
     forEach(['addressLine1', 'addressLine2', 'addressCity', 'addressPostcode'])
-    .it('address fields should not exist', field => {
+    .it.skip('address fields should not exist', field => {
         // Arrange
         DemoControls.selectCheckoutMethod('collection');
 
@@ -24,12 +29,11 @@ describe('f-checkout "collection" component tests', () => {
         expect(CheckoutComponent.doesInputFieldExist(field)).toBe(false);
     });
 
-    forEach(['mobileNumber'])
-    .it('should display the mandatory fields', field => {
+    it.skip('should display the mandatory "mobileNumber" field', () => {
         // Arrange
         DemoControls.selectCheckoutMethod('collection');
 
         // Assert
-        expect(CheckoutComponent.isFieldDisplayed(field)).toBe(true);
+        expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
     });
 });