+++ packages/f-registration/src/components/Registration.vue
@@ -18,10 +18,10 @@
                 ]"
                 data-test-id="create-account-login-link"
                 @click="visitLoginPage">
-                <a :href="copy.navLinks.login.url">{{ copy.navLinks.login.text }}</a>
+                <a :href="loginUrl">{{ copy.navLinks.login.text }}</a>
             </p>
             <form
-                type="post"
+                method="post"
                 :class="$style['c-registration-form']"
                 @click="formStart"
                 @focus="formStart"
@@ -244,7 +244,11 @@ export default {
         },
         showLoginLink: {
             type: Boolean,
-            default: true
+            required: true
+        },
+        loginUrl: {
+            type: String,
+            required: true
         }
     },