+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v2.24.0
+------------------------------
+*December 11, 2020*
+
+### Changed
+- package.json component test scripts
+
+
+v2.23.0
+------------------------------
+*December 9, 2020*
+
+### Changed
+- CircleCI config to build / serve Storybook for component tests
+- package.json component test scripts
+
+### Removed
+- Component demo files
+
+
 v2.22.1
 ------------------------------
 *December 8, 2020*