+++ packages/f-checkout/CHANGELOG.md
@@ -3,6 +3,40 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.22.0
+------------------------------
+*December 11, 2020*
+
+### Added
+- Webdriver tests for postcode format in `f-checkout-delivery`.
+
+
+v0.21.0
+------------------------------
+*December 10, 2020*
+
+### Added
+- Webdriver test for phone number format in `f-checkout-delivery`.
+
+
+v0.20.1
+------------------------------
+*December 10, 2020*
+
+### Changed
+- `f-form-field` version.
+- Styling of `Address` fields in error.
+
+
+v0.20.0
+------------------------------
+*December 9, 2020*
+
+### Changed
+- `f-form-field` version.
+- Styling in `Address` component.
+
+
 v0.19.0
 ------------------------------
 *December 8, 2020*