+++ packages/f-registration/CHANGELOG.md
@@ -3,6 +3,29 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.44.1
+------------------------------
+*December 10, 2020*
+
+### Fixed
+- Set form method to POST.
+
+
+v0.44.0
+------------------------------
+*December 9, 2020*
+
+### Added
+- 'loginUrl' prop so that returnUrl is not lost when clicking 'Already on Just Eat' as it can be passed in by the parent
+
+### Changed
+- Updated storybook to include missing props
+- Make 'showLoginLink' a required prop
+
+### Removed
+- Unsupported EU tenants
+- Obselete localisation message for loginUrl
+
 
 v0.43.3
 ------------------------------