+++ packages/f-vue-icons/CHANGELOG.md
@@ -3,6 +3,51 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+
+v2.0.1
+------------------------------
+*December 9, 2020*
+
+### Fixed
+- Question mark icon filename casing.
+
+
+v2.0.0
+------------------------------
+*December 9, 2020*
+
+### Added
+- Arrow icon.
+- Celebrate bag bg icon.
+- Bag icon.
+- Bin icon.
+- Chevron split icon.
+- Moped icon.
+
+## Changed
+- Updated npm dependencies.
+- Updated alert icon.
+- Updated celebrate bag icon.
+- Updated basket icon.
+- Updated chevron icon.
+- Updated cross icon.
+- Updated info icon.
+- Updated minus icon.
+- Updated offer icon.
+- Updated plus icon.
+- Updated spicy icon.
+- Updated star icons.
+- Updated vegetarian icon.
+- Updated warning icon.
+- Renamed question mark icon.
+
+### Removed
+- Collection icons (use bag icons instead).
+- Delivery icon (use moped icon instead).
+- Legacy JE logo.
+- Plus thick icon.
+
+
 v1.13.0
 ------------------------------
 *December 1, 2020*